Changed variables data types accordingly, and fixed casting exceptions

+++ src/main/java/io/ipgeolocation/api/GeolocationTimezone.java
@@ -1,29 +1,30 @@
 package io.ipgeolocation.api;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public class GeolocationTimezone {
     private final String name;
-    private final Double offset;
+    private final Integer offset;
     private final String currentTime;
-    private final Double currentTimeUnix;
+    private final BigDecimal currentTimeUnix;
     private final Boolean isDST;
-    private final Double dstSavings;
+    private final Integer dstSavings;
 
     GeolocationTimezone(Map<String, Object> json) {
         this.name = (String) json.get("name");
-        this.offset = (Double) json.get("offset");
+        this.offset = (Integer) json.get("offset");
         this.currentTime = (String) json.get("current_time");
-        this.currentTimeUnix = (Double) json.get("current_time_unix");
+        this.currentTimeUnix = (BigDecimal) json.get("current_time_unix");
         this.isDST = (Boolean) json.get("is_dst");
-        this.dstSavings = (Double) json.get("dst_savings");
+        this.dstSavings = (Integer) json.get("dst_savings");
     }
 
     public String getName() {
         return name;
     }
 
-    public Double getOffset() {
+    public Integer getOffset() {
         return offset;
     }
 
@@ -31,7 +32,7 @@ public class GeolocationTimezone {
         return currentTime;
     }
 
-    public Double getCurrentTimeUnix() {
+    public BigDecimal getCurrentTimeUnix() {
         return currentTimeUnix;
     }
 
@@ -39,7 +40,7 @@ public class GeolocationTimezone {
         return isDST;
     }
 
-    public Double getDSTSavings() {
+    public Integer getDSTSavings() {
         return dstSavings;
     }
